LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/form-roles-expected.txt

Verifies 9.1 Roles - handling author errors and the form role.

x
x

FAIL form without label assert_equals: <nav role="form" data-testname="form without label" data-expectedrole="navigation" class="ex">x</nav> expected "navigation" but got "form"
PASS form with label
